# Parallelize detensorizing token IDs during tree search