@stephenroller 2021-06-16
do we need to port this change to torchscript?