@EricMichaelSmith 2021-06-17
> do we need to port this change to torchscript?

Hmm, doesn't look like it, because torchscript export currently only supports greedy search